Work around email bug when admins are BCCed.
When admins are BCCed on AppEngine, all other receipients appear to be
removed, and the admins are promoted to the TO field of the email.
The fix is to send a separated email with the admins in the to field.
This also fixes the test stub so that recipients can be retrieved from
emails, at the cost of having to filter out duplicates in the integration
test helper.